refactor(release): separate PDFium archive responsibilities
