+++ crates/pdfium/src/native.rs
@@ -15,6 +15,9 @@ use std::io::{Read as _, Write as _};
 use std::path::{Path, PathBuf};
 use std::sync::atomic::{AtomicBool, Ordering};
 
+mod runtime_path;
+use runtime_path::contains_link_or_reparse;
+
 struct FixedOutput<T> {
     slots: Box<[std::mem::MaybeUninit<T>]>,
     initialized: usize,
@@ -1977,29 +1980,6 @@ fn validated_snapshot(path: &Path, artifact: &Artifact) -> Result<Snapshot, Erro
     validated_snapshot_with_hook(path, artifact, || {}, || {})
 }
 
-fn path_contains_link_or_reparse(path: &Path) -> Result<bool, Error> {
-    let mut current = PathBuf::new();
-    for component in path.components() {
-        current.push(component.as_os_str());
-        if matches!(component, std::path::Component::Prefix(_)) {
-            continue;
-        }
-        let metadata = fs::symlink_metadata(&current)
-            .map_err(|error| Error::InvalidPath(error.to_string()))?;
-        if metadata.file_type().is_symlink() {
-            return Ok(true);
-        }
-        #[cfg(windows)]
-        {
-            use std::os::windows::fs::MetadataExt as _;
-            if metadata.file_attributes() & 0x400 != 0 {
-                return Ok(true);
-            }
-        }
-    }
-    Ok(false)
-}
-
 #[allow(clippy::too_many_lines)] // Keep the security-sensitive open, hash, copy, and lock sequence linear.
 fn validated_snapshot_with_hook(
     path: &Path,
@@ -2014,7 +1994,7 @@ fn validated_snapshot_with_hook(
             "path must be absolute and use the pinned library filename".into(),
         ));
     }
-    if path_contains_link_or_reparse(path)? {
+    if contains_link_or_reparse(path)? {
         return Err(Error::InvalidPath(
             "runtime path must not contain a symbolic link or reparse point".into(),
         ));
@@ -2461,66 +2441,6 @@ mod tests {
         let path = dir.path().join(&artifact.library);
         File::create(&path).unwrap().set_len(artifact.library_size).unwrap();
         assert!(matches!(Native::load(&path), Err(Error::DigestMismatch { .. })));
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn explicit_runtime_rejects_linked_file_and_parent_directory() {
-        use std::os::unix::fs::symlink;
-
-        let dir = tempfile::tempdir().unwrap();
-        let outside = dir.path().join("outside");
-        fs::create_dir(&outside).unwrap();
-        let artifact = Platform::current().unwrap().artifact().unwrap();
-        let source = outside.join(&artifact.library);
-        fs::write(&source, b"good").unwrap();
-
-        let linked_file = dir.path().join(&artifact.library);
-        symlink(&source, &linked_file).unwrap();
-        assert!(matches!(
-            crate::Pdfium::load_pinned(&linked_file, Limits::default()),
-            Err(Error::InvalidPath(message)) if message.contains("symbolic link or reparse point")
-        ));
-
-        let linked_parent = dir.path().join("runtime");
-        symlink(&outside, &linked_parent).unwrap();
-        assert!(matches!(
-            crate::Pdfium::load_pinned(
-                &linked_parent.join(&artifact.library),
-                Limits::default(),
-            ),
-            Err(Error::InvalidPath(message)) if message.contains("symbolic link or reparse point")
-        ));
-    }
-
-    #[cfg(windows)]
-    #[test]
-    fn explicit_runtime_rejects_linked_file_and_reparse_parent_directory() {
-        use std::os::windows::fs::{symlink_dir, symlink_file};
-
-        let dir = tempfile::tempdir().unwrap();
-        let outside = dir.path().join("outside");
-        fs::create_dir(&outside).unwrap();
-        let artifact = Platform::current().unwrap().artifact().unwrap();
-        let source = outside.join(&artifact.library);
-        fs::write(&source, b"good").unwrap();
-
-        let linked_file = dir.path().join(&artifact.library);
-        symlink_file(&source, &linked_file).unwrap();
-        assert!(matches!(
-            crate::Pdfium::load_pinned(&linked_file, Limits::default()),
-            Err(Error::InvalidPath(message)) if message.contains("symbolic link or reparse point")
-        ));
-
-        let linked_parent = dir.path().join("runtime");
-        symlink_dir(&outside, &linked_parent).unwrap();
-        assert!(matches!(
-            crate::Pdfium::load_pinned(
-                &linked_parent.join(&artifact.library),
-                Limits::default(),
-            ),
-            Err(Error::InvalidPath(message)) if message.contains("symbolic link or reparse point")
-        ));
     }
 
     #[test]
