+++ tools/portable-release/BUILD.bazel
@@ -1,4 +1,7 @@
 exports_files([
     "assemble.py",
+    "core_archive.py",
     "native_acceptance.py",
+    "post_release_scenarios.py",
+    "release_artifacts.py",
 ])
